requirements.txt
torch==2.0.1
torchvision==0.15.2
onnx
pillow==10.2.0
IPython
gdown
tqdm
